move client pool to new location
